web/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# web/Application/__init__.py: 1
# web/Application/auth.py: 1
# web/Application/dashboard.py: 2
# web/Migrations/env.py: 6
Flask == 1.1.2

# web/Application/__init__.py: 3
# web/Application/auth.py: 2
# web/Application/dashboard.py: 3
# web/Application/models.py: 2
Flask_Login == 0.5.0

# web/Application/__init__.py: 4
Flask_Migrate == 2.7.0

# web/Application/__init__.py: 2
Flask_SQLAlchemy == 2.4.4

# web/Application/forms.py: 7
Flask_WTF == 0.14.3

# web/Migrations/versions/2be8e017b4a8_file_detail_model.py: 9
# web/Migrations/versions/f1c24864ca94_user_model.py: 9
SQLAlchemy == 1.3.23

# web/Application/forms.py: 4,5,10,11
WTForms == 2.3.3

# web/Application/dashboard.py: 8
# web/Application/forms.py: 8
# web/Application/models.py: 3
Werkzeug == 1.0.1

# web/Migrations/env.py: 8
# web/Migrations/versions/2be8e017b4a8_file_detail_model.py: 8
# web/Migrations/versions/f1c24864ca94_user_model.py: 8
alembic == 1.5.6

# web/config.py: 2
python_dotenv == 0.15.0

# web/Application/dashboard.py: 15
requests == 2.25.1
